+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 * Features:
   * Display encumbrance value for gear items in the material tab
-  * Add the icon to purchase skill next the skill name
-  * Add settings to add an icon to display the skill description from compendium next the skill name
+  * Add the icon next the skill name to purchase skill
+  * Add settings to add an icon next the skill name to open the skill description from compendium
+  * Add setting to select the compendium where find skills description
 * Fixes:
   * Award icon next to skill name can be use to increase the skill rank
   * Display adjusted and value damage for weapons